rename routes folder to auth folder
