+++ frontend/src/App.js
@@ -1,12 +1,12 @@
 import React, { Fragment } from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import './styles/App.css';
 import Alerts from './components/layout/Alerts';
 import ClientApp from './components/client/ClientApp';
-import Login from './components/routes/Login';
+import Login from './components/auth/Login';
 import Navbar from './components/layout/Navbar';
-import PrivateRoute from './components/routes/PrivateRoute';
-import Register from './components/routes/Register';
+import PrivateRoute from './components/auth/PrivateRoute';
+import Register from './components/auth/Register';
 import AlertState from './context/alert/AlertState';
 import AuthState from './context/auth/AuthState';
 import ClientState from './context/client/ClientState';
@@ -17,7 +17,7 @@ const App = () => {
       <AuthState>
         <ClientState>
           <AlertState>
-            <BrowserRouter>
+            <Router>
               <Fragment>
                 <Navbar />
                 <Alerts />
@@ -32,7 +32,7 @@ const App = () => {
                   <PrivateRoute component={ClientApp} />
                 </Switch>
               </Fragment>
-            </BrowserRouter>
+            </Router>
           </AlertState>
         </ClientState>
       </AuthState>
